part main function moved to server
